updated structure, ran local tests, and compiled doc
