+++ docs/info/VISDA.md
@@ -8,4 +8,4 @@
 | Bias                | 100.0 $\mathrm{DN}$                           |
 | Wavelength Midpoint | 0.56 $\mathrm{\mu m}$                         |
 | Integration Time    | 0.2 $\mathrm{s}$                              |
-| Zeropoint           | 3.765e-09$\mathrm{\frac{erg}{A\,s\,cm^{2}}}$  |
+| Zeropoint           | 3.768e-09$\mathrm{\frac{erg}{A\,s\,cm^{2}}}$  |
