+++ pyproject.toml
@@ -5,21 +5,21 @@ description = ""
 authors = ["Christina Hedges <christina.l.hedges@nasa.gov>"]
 
 [tool.poetry.dependencies]
-python = ">=3.10,<3.13"
+python = ">=3.9,<3.13"
 numpy = ">=1.26.0"
 astropy = ">=5.3.3"
 matplotlib = ">=3.8.0"
 scipy = ">=1.11.2"
 tqdm = ">=4.60.1"
 rich = ">=13.0.0"
-synphot = ">=1.5.0"
-stsynphot = ">=1.4.0"
 astroquery = ">=0.4.6"
 pyarrow = "<20.0.0"
 openpyxl = "^3.1.5"
 appdirs = "^1.4.4"
 pandas = {version = "^2.2.3", extras = ["tabulate"]}
 tabulate = "^0.9.0"
+synphot = ">=1.3.0"
+stsynphot = ">=1.3.0"
 
 [tool.poetry.group.dev]
 optional = true
