+++ src/pandorasat/utils.py
@@ -1,5 +1,4 @@
 # Standard library
-import os
 import warnings
 from functools import lru_cache
 
@@ -14,12 +13,11 @@ from astropy.time import Time
 from astroquery.gaia import Gaia
 
 from . import PACKAGEDIR, __version__
+from .phoenix import get_phoenix_model
 
 
 def SED(teff, logg=4.5, jmag=None, vmag=None):
     """Gives a model SED for a given Teff, logg and magnitude."""
-    from .phoenix import get_phoenix_model
-
     return get_phoenix_model(teff, logg=logg, jmag=jmag, vmag=vmag)
 
 
@@ -185,22 +183,6 @@ def simulate_flatfield(stddev=0.005, seed=777):
             checksum=True,
         )
     return
-
-
-def load_vega():
-    from .phoenix import PHOENIXPATH, get_vega
-
-    os.environ["PYSYN_CDBS"] = PHOENIXPATH
-    get_vega()
-    # Third-party
-    import stsynphot  # noqa : F401
-    from synphot import SourceSpectrum
-
-    vega = SourceSpectrum.from_vega()
-    wavelength, spectrum = vega.waveset, vega(vega.waveset, flux_unit="flam")
-
-    spectrum = spectrum.value * u.erg / u.cm**2 / u.s / u.angstrom
-    return wavelength, spectrum
 
 
 def load_benchmark():
