+++ tests/test_phoenix.py
@@ -16,8 +16,22 @@ import stsynphot as stsyn  # noqa: E402
 
 # testing get_vega
 def test_get_vega():
-    phoenix.get_vega()
+    if os.getenv("GITHUB_ACTIONS") == "true":
+        pytest.skip(
+            "Skipping this test on GitHub Actions this downloads a database of stellar models."
+        )
+
+    phoenix.download_vega()
     assert os.path.isfile(PHOENIXPATH + "/calspec/alpha_lyr_stis_011.fits")
+
+    wav, spec = phoenix.load_vega()
+
+    # Check that the loaded file is correct
+    assert isinstance(wav, u.Quantity)
+    assert isinstance(spec, u.Quantity)
+    assert len(wav) > 0
+    assert len(spec) > 0
+    return
 
 
 # test get_phoenix_model
